Change CLI to have similar interface as javafx so they would be interchangeable
Also fixed a bug with Cli.split method.
