+++ src/main/java/io/Ui.java
@@ -9,15 +9,4 @@ public interface Ui {
      * @param reply Message to reply with
      */
     public void showReply(String reply);
-
-    /**
-     * Starts the ui and performs any preliminary actions
-     */
-    public Ui launch();
-
-    /**
-     * Obtains user input from the ui
-     * @return User input
-     */
-    public String getInput();    
 }
